Move ruamel.yaml dependencies to interface_py_c.py
