+++ svmbir/_utils.py
@@ -7,86 +7,6 @@ import warnings
 import hashlib
 # pdb.set_trace()
 
-from ruamel.yaml import YAML
-
-
-##################################
-## mbir read/modify Param Files ##
-##################################
-
-def parse_params(default_params, **kwargs):
-    params = dict(default_params)
-    common_keys = set(kwargs.keys()) & set(params.keys())
-    for key in common_keys :
-        params[key] = kwargs[key]
-
-    return params
-
-
-def read_params(params_path):
-    with open(params_path, 'r') as fileID :
-        yaml = YAML()
-        params = yaml.load(fileID)
-
-    return params
-
-
-def print_params(params, start_str = ''):
-    for key, value in params.items() :
-        if isinstance(value, dict) :
-            print('{}:'.format(key))
-            print_params(value, start_str='    ')
-        else :
-            print(start_str + '{}: {}'.format(key, value))
-
-
-def modify_params(filePath, **kwargs):
-    with open(filePath, 'r') as fileID :
-        yaml = YAML()
-        yaml_dict = yaml.load(fileID)
-
-    # print(kwargs.keys())
-
-    for key in kwargs.keys() :
-        yaml_dict[key] = kwargs[key]
-
-    with open(filePath, 'w') as fileID :
-        yaml.dump(yaml_dict, fileID)
-
-
-def sanitize_params(params):
-    if isinstance(params, dict):
-        params = dict(params)
-        for key in params:
-            params[key] = sanitize_params(params[key])
-
-    if isinstance(params, (np.ndarray, np.generic)):
-        params = params.tolist()
-
-    return params
-
-
-def write_params(filePath, **kwargs):
-    kwargs = sanitize_params(kwargs)
-    # print(kwargs)
-    # sys.stdout.flush()
-
-    with open(filePath, 'w') as fileID :
-        yaml = YAML()
-        yaml.dump(kwargs, fileID)
-
-
-def readAngleList(filePath):
-    with open(filePath, 'r') as fileID :
-        lines = fileID.read().split("\n")
-
-    angleList = []
-    for line in lines :
-        if not line.isspace() and line :
-            angleList.append(float(line))
-
-    return angleList
-
 
 #########################################
 ## mbir read/write/delete Binary Files ##
